Two constructors which are used to initialize new instances (objects) of the class with specific values.
****Note missing public class from the second construct public Employee(String name, String email)
this.name = "Default Name";/this.email = "default@example.com"; setting name to default name and email, this.empNum = nextEmpNum++; assigning unique number to employee to the static field nextEmpNum and then incrementing this field for the next employee. Second set of constructs will allow user to create custom values for name and email this. keyword lets us to refer to current object. Also I have seen that it will work out better to not split first and last name keep it as just name.