Email validation Regex
I have created Regex so that "The value must not be accepted if its length is 3 or less.". If its not valid error message will be printed. Changed this.email = email; to setEmail(email);. This will ensure that any email in the code will have to pass the validation test (Regex)