Search and display the name
I have created a simple for each loop, that goes over each employee in the array, if statement asks program to only display employees who have an employee number above the value stored in a variable called m. But when testing the first names are being duplicated for each employee.