Fixing search and display employees
Soo to fix this i have created/overrided toString in Employee class so that it displays information in the correct format beside the name we have email address and number of each of our employees in the array. In the EmployeeTest I have changed the system print out from EmployeeTest class where I had issues as the first names were being duplicated, to read information from created string in Employee class where all of the info is being stored. I feel like its more convenient as we can see in that class what employee details each of them has and which we want to display in the terminal.