Manager class username/password
Manager class that extends Employee class, where we are adding unique fields for manager which are username and password. We are calling superclass Employee constructor with name and email these fields are initialised in line with employee object requirements. Then we are initialising username and password for the Manger object. We share structor of an normal employee but adding manager like features username and password etc.