Setting up manager console for diffrent operations
Scanner has been created for the users input of the correct username and password to be able to run different options that manager has when it comes to managing employees. If username and password are true program will set running flag which will allow manager to keep menu loop running until they will decide to exit.